Refactor backend: Modularize main.py into structured architecture
