+++ backend/exceptions/handlers.py
@@ -1,0 +1,29 @@
+from fastapi import Request, HTTPException
+from fastapi.responses import JSONResponse
+from datetime import datetime
+import logging
+
+logger = logging.getLogger(__name__)
+
+async def http_exception_handler(request: Request, exc: HTTPException):
+    return JSONResponse(
+        status_code=exc.status_code,
+        content={
+            "error": True,
+            "message": exc.detail,
+            "timestamp": datetime.now().isoformat(),
+            "path": str(request.url)
+        }
+    )
+
+async def general_exception_handler(request: Request, exc: Exception):
+    logger.error(f"Unexpected error: {str(exc)}")
+    return JSONResponse(
+        status_code=500,
+        content={
+            "error": True,
+            "message": "Internal server error",
+            "timestamp": datetime.now().isoformat(),
+            "path": str(request.url)
+        }
+    )
